pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "memora"
version = "0.1.0"
description = "Local-first decision intelligence platform"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Memora Team" }]

dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "duckdb>=1.0",
    "weaviate-client>=4.0",
    "sentence-transformers>=3.0",
    "openai>=1.60",
    "langgraph>=0.2",
    "apscheduler>=3.10",
    "httpx>=0.27",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.8",
]

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py312"
line-length = 100
